+++ app/src/components/market/common/display_arbitrator/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 
 import { Arbitrator } from '../../../../util/types'
-import { ArbitratorIcon } from '../arbitrator_icon'
+import { ArbitratorIcon } from '../../sections/common_sections/arbitrators/arbitrator_icon'
 
 const AWrapper = styled.a`
   text-decoration: none !important;
